Refactor cmdutil.NewRepo back into ghrepo as FromName
